change module name from androidsdk to remotesdk - move all files to new folder & packaging them - change name of JalaliEventsDb to RemoteJalaliEventsDb class - change name of CalendarEvents to RemoteCalendarEvents class
