--- androidsdk/objectbox-models/default.json.bak
@@ -1,81 +1,0 @@
-{
-  "_note1": "KEEP THIS FILE! Check it into a version control system (VCS) like git.",
-  "_note2": "ObjectBox manages crucial IDs for your object model. See docs for details.",
-  "_note3": "If you have VCS merge conflicts, you must resolve them according to ObjectBox docs.",
-  "entities": [
-    {
-      "id": "1:5530299652600405768",
-      "lastPropertyId": "10:8280567662530696932",
-      "name": "JalaliDb",
-      "properties": [
-        {
-          "id": "1:6552609070583710525",
-          "name": "calendar",
-          "type": 9
-        },
-        {
-          "id": "2:2841101781713711304",
-          "name": "month",
-          "type": 5,
-          "flags": 2
-        },
-        {
-          "id": "3:8441957179738781562",
-          "name": "sources",
-          "type": 9
-        },
-        {
-          "id": "4:1397163860474224995",
-          "name": "year",
-          "type": 5,
-          "flags": 2
-        },
-        {
-          "id": "5:5643902470924683833",
-          "name": "description_fa_IR",
-          "type": 9
-        },
-        {
-          "id": "6:6240750917548808505",
-          "name": "title_fa_IR",
-          "type": 9
-        },
-        {
-          "id": "7:1331539828783851801",
-          "name": "day",
-          "type": 5,
-          "flags": 2
-        },
-        {
-          "id": "8:5732093623234541236",
-          "name": "holiday_Iran",
-          "type": 9
-        },
-        {
-          "id": "9:4657296545268146423",
-          "name": "key",
-          "type": 6,
-          "flags": 2
-        },
-        {
-          "id": "10:8280567662530696932",
-          "name": "id",
-          "type": 6,
-          "flags": 1
-        }
-      ],
-      "relations": []
-    }
-  ],
-  "lastEntityId": "1:5530299652600405768",
-  "lastIndexId": "0:0",
-  "lastRelationId": "0:0",
-  "lastSequenceId": "0:0",
-  "modelVersion": 5,
-  "modelVersionParserMinimum": 5,
-  "retiredEntityUids": [],
-  "retiredIndexUids": [],
-  "retiredPropertyUids": [],
-  "retiredRelationUids": [],
-  "version": 1
-}
