--- androidsdk/src/main/java/com/github/persiancal/androidsdk/model/jalali/Description.kt
@@ -1,9 +1,0 @@
-package com.github.persiancal.androidsdk.model.jalali
-
-import com.google.gson.annotations.SerializedName
-
-data class Description(
-
-    @field:SerializedName("fa_IR")
-    val faIR: String? = null
-)
