--- androidsdk/src/main/java/com/github/persiancal/androidsdk/model/jalali/Holiday.kt
@@ -1,9 +1,0 @@
-package com.github.persiancal.androidsdk.model.jalali
-
-import com.google.gson.annotations.SerializedName
-
-data class Holiday(
-
-    @field:SerializedName("Iran")
-    val iran: List<String>? = null
-)
