--- androidsdk/src/main/java/com/github/persiancal/androidsdk/model/jalali/JalaliEventsDb.kt
@@ -1,24 +1,0 @@
-package com.github.persiancal.androidsdk.model.jalali
-
-import com.github.persiancal.androidsdk.util.StringListConverter
-import io.objectbox.annotation.Convert
-import io.objectbox.annotation.Entity
-import io.objectbox.annotation.Id
-
-@Entity
-data class JalaliEventsDb(
-    @Id
-    var id: Long = 0,
-    val key: Long?,
-    @Convert(converter = StringListConverter::class, dbType = String::class)
-    val calendar: List<String?>? = null,
-    val month: Int? = null,
-    @Convert(converter = StringListConverter::class, dbType = String::class)
-    val sources: List<String?>? = null,
-    val year: Int? = null,
-    val description_fa_IR: String? = null,
-    val title_fa_IR: String? = null,
-    val day: Int? = null,
-    @Convert(converter = StringListConverter::class, dbType = String::class)
-    val holiday_Iran: List<String>? = null
-)
