--- androidsdk/src/main/java/com/github/persiancal/androidsdk/model/jalali/NameItem.kt
@@ -1,12 +1,0 @@
-package com.github.persiancal.androidsdk.model.jalali
-
-import com.google.gson.annotations.SerializedName
-
-data class NameItem(
-
-    @field:SerializedName("en_US")
-    val enUS: String? = null,
-
-    @field:SerializedName("fa_IR")
-    val faIR: String? = null
-)
