--- androidsdk/src/main/java/com/github/persiancal/androidsdk/util/Constants.kt
@@ -1,7 +1,0 @@
-package com.github.persiancal.androidsdk.util
-
-class Constants {
-    companion object {
-        const val BASE_URL = "https://raw.githubusercontent.com/persiancal/cal-events/gh-pages/"
-    }
-}
