+++ app/src/main/java/com/github/persiancal/android/MyApplication.kt
@@ -2,12 +2,12 @@ package com.github.persiancal.android
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
-import com.github.persiancal.androidsdk.CalendarEvents
+import com.github.persiancal.sdkremote.RemoteCalendarEvents
 
 class MyApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-        CalendarEvents.init(this)
+        RemoteCalendarEvents.init(this)
     }
 }
