+++ sdk-remote/objectbox-models/default.json
@@ -1,0 +1,105 @@
+{
+  "_note1": "KEEP THIS FILE! Check it into a version control system (VCS) like git.",
+  "_note2": "ObjectBox manages crucial IDs for your object model. See docs for details.",
+  "_note3": "If you have VCS merge conflicts, you must resolve them according to ObjectBox docs.",
+  "entities": [
+    {
+      "id": "3:455104927502078738",
+      "lastPropertyId": "10:5009066921513924187",
+      "name": "RemoteJalaliEventsDb",
+      "properties": [
+        {
+          "id": "1:6039848384105223855",
+          "name": "id",
+          "type": 6,
+          "flags": 1
+        },
+        {
+          "id": "2:240562119042028611",
+          "name": "key",
+          "type": 6,
+          "flags": 2
+        },
+        {
+          "id": "3:6001868522676492561",
+          "name": "calendar",
+          "type": 9
+        },
+        {
+          "id": "4:6713982720090624006",
+          "name": "month",
+          "type": 5,
+          "flags": 2
+        },
+        {
+          "id": "5:5125634243503002018",
+          "name": "sources",
+          "type": 9
+        },
+        {
+          "id": "6:2243168573978315827",
+          "name": "year",
+          "type": 5,
+          "flags": 2
+        },
+        {
+          "id": "7:2978321448288420467",
+          "name": "description_fa_IR",
+          "type": 9
+        },
+        {
+          "id": "8:6661719221686178747",
+          "name": "title_fa_IR",
+          "type": 9
+        },
+        {
+          "id": "9:6532227170239232748",
+          "name": "day",
+          "type": 5,
+          "flags": 2
+        },
+        {
+          "id": "10:5009066921513924187",
+          "name": "holiday_Iran",
+          "type": 9
+        }
+      ],
+      "relations": []
+    }
+  ],
+  "lastEntityId": "3:455104927502078738",
+  "lastIndexId": "0:0",
+  "lastRelationId": "0:0",
+  "lastSequenceId": "0:0",
+  "modelVersion": 5,
+  "modelVersionParserMinimum": 5,
+  "retiredEntityUids": [
+    5530299652600405768,
+    5707509275168758776
+  ],
+  "retiredIndexUids": [],
+  "retiredPropertyUids": [
+    6552609070583710525,
+    2841101781713711304,
+    8441957179738781562,
+    1397163860474224995,
+    5643902470924683833,
+    6240750917548808505,
+    1331539828783851801,
+    5732093623234541236,
+    4657296545268146423,
+    8280567662530696932,
+    3777039416076769967,
+    7653564742772940671,
+    3233422486222122358,
+    1080416884604312582,
+    4090982089483201327,
+    5804187689365615862,
+    628607950202485544,
+    608930713493712089,
+    834096886594308873,
+    6653753579581522335
+  ],
+  "retiredRelationUids": [],
+  "version": 1
+}
