+++ sdk-remote/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.github.persiancal.sdkremote" />
