+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/jalali/CalendarsItem.kt
@@ -1,0 +1,12 @@
+package com.github.persiancal.sdkremote.model.jalali
+
+import com.google.gson.annotations.SerializedName
+
+data class CalendarsItem(
+
+    @field:SerializedName("en_US")
+    val enUS: String? = null,
+
+    @field:SerializedName("fa_IR")
+    val faIR: String? = null
+)
