+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/jalali/Description.kt
@@ -1,0 +1,9 @@
+package com.github.persiancal.sdkremote.model.jalali
+
+import com.google.gson.annotations.SerializedName
+
+data class Description(
+
+    @field:SerializedName("fa_IR")
+    val faIR: String? = null
+)
