+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/jalali/Holiday.kt
@@ -1,0 +1,9 @@
+package com.github.persiancal.sdkremote.model.jalali
+
+import com.google.gson.annotations.SerializedName
+
+data class Holiday(
+
+    @field:SerializedName("Iran")
+    val iran: List<String>? = null
+)
