+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/jalali/Months.kt
@@ -1,0 +1,15 @@
+package com.github.persiancal.sdkremote.model.jalali
+
+import com.google.gson.annotations.SerializedName
+
+data class Months(
+
+    @field:SerializedName("normal")
+    val normal: List<Int?>? = null,
+
+    @field:SerializedName("name")
+    val name: List<NameItem?>? = null,
+
+    @field:SerializedName("leap")
+    val leap: List<Int?>? = null
+)
