+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/model/jalali/NameItem.kt
@@ -1,0 +1,12 @@
+package com.github.persiancal.sdkremote.model.jalali
+
+import com.google.gson.annotations.SerializedName
+
+data class NameItem(
+
+    @field:SerializedName("en_US")
+    val enUS: String? = null,
+
+    @field:SerializedName("fa_IR")
+    val faIR: String? = null
+)
