+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/service/ApiService.kt
@@ -1,0 +1,12 @@
+package com.github.persiancal.sdkremote.service
+
+import com.github.persiancal.sdkremote.model.jalali.JalaliResponse
+import io.reactivex.Single
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface ApiService {
+
+    @GET("{endpoint}")
+    fun getJalaliEvents(@Path("endpoint") get: String): Single<JalaliResponse>
+}
