+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/util/Constants.kt
@@ -1,0 +1,7 @@
+package com.github.persiancal.sdkremote.util
+
+class Constants {
+    companion object {
+        const val BASE_URL = "https://raw.githubusercontent.com/persiancal/cal-events/gh-pages/"
+    }
+}
