+++ sdk-remote/src/main/java/com/github/persiancal/sdkremote/util/StringListConverter.kt
@@ -1,0 +1,23 @@
+package com.github.persiancal.sdkremote.util
+
+
+import io.objectbox.converter.PropertyConverter
+
+class StringListConverter : PropertyConverter<List<String>, String> {
+    override fun convertToEntityProperty(databaseValue: String?): List<String> {
+        if (databaseValue == null) return ArrayList()
+        return databaseValue.split(",")
+    }
+
+    override fun convertToDatabaseValue(entityProperty: List<String>?): String {
+        if (entityProperty == null) return ""
+        if (entityProperty.isEmpty()) return ""
+        if (entityProperty.size == 1) {
+            return entityProperty[0]
+        }
+        val builder = StringBuilder()
+        entityProperty.forEach { builder.append(it).append(",") }
+        builder.deleteCharAt(builder.length - 1)
+        return builder.toString()
+    }
+}
