+++ sdk-remote/src/main/res/values/strings.xml
